EAD Session-8 (Spring employee portal) Assignment

+++ EAD_Assignments/Spring_Assignments/EADSession_8/EmployeePortal/src/main/java/com/metacube/training/EmployeePortal/service/ProjectService.java
@@ -8,11 +8,11 @@ import com.metacube.training.EmployeePortal.model.Project;
 public interface ProjectService {
 
 	
-	Project getProjectById(Long id);
+	Project getProjectById(int id);
 
 	List<Project> getAllProjects();
 
-	boolean deleteProject(Long id);
+	boolean deleteProject(int id);
 
 	boolean updateProject(Project project);
